+++ common/routes/Home/components/VideoBlock.js
@@ -25,8 +25,9 @@ class VideoBlock extends PureComponent {
 
   onSelectThumb (index) {
     this.setState({selectedIndex: index});
-    // 2330px = VideoBlock 到底下的高度 = document.body.scrollHeight - document.body.scrollTop
-    window.scroll(0, document.body.scrollHeight - 2330);
+    // 70 是 margin-top 要額外減掉
+    var videoOffsetTop = document.getElementById('home-video-block').offsetTop;
+    window.scroll(0, videoOffsetTop - 70);
   }
 
   prev () {
@@ -54,8 +55,10 @@ class VideoBlock extends PureComponent {
 
     return (
       <Margin10 className='relative'>
-        <div key={this.state.selectedIndex}>
-          <VideoPlayer height={500} poster={selectedItem.MainPhoto.large} src={selectedItem.MainVideo.url} />
+        <div className={css(styles.videoPlayer)}>
+          <div key={this.state.selectedIndex}>
+            <VideoPlayer height={500} poster={selectedItem.MainPhoto.large} src={selectedItem.MainVideo.url} />
+          </div>
         </div>
         <FontAwesome className={css(styles.iconLeft)} name='chevron-left' size='2x'
           onClick={this.prev} />
@@ -127,6 +130,11 @@ const styles = StyleSheet.create({
     background: '#4A4C4D',
     height: 55.5,
     padding: '5px 10px'
+  },
+  videoPlayer: {
+    background: '#000000',
+    width: 970,
+    height: 500
   }
 });