+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Changed
 - 精選特輯換成火線話題 @esbb48
 - 對調專題與特輯的位置 @esbb48
+- 微調首頁 Video 的樣式，讓換影片的體驗不會有瞬間抽換的感覺 @esbb48
 
 ### Changed
 - 使用新版 imgLib  @esbb48